# openapi.yaml
openapi: 3.1.0
info:
  title: cam-sim control API
  version: '1'
  description: >
    Controls a simulated Reolink RLC-1224A camera: events, recordings, faults,
    actions and state. Every /sim/api route needs `Authorization: Bearer <CAMSIM_CONTROL_TOKEN>`,
    or the web UI's session cookie (from /sim/login); with the cookie, writes also need the
    header `X-CamSim-UI: 1`. A token in the URL is refused. Without a configured token the API
    answers 404.
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        detail: { type: string }
    Fault:
      type: object
      required: [name]
      properties:
        name:
          type: string
          enum: [downloads.refuse, downloads.dropFirst, downloads.dropMidway, downloads.delayMs, flv.reset, flv.delayMs, search.delayMs, settings.fail, settings.ignore, settings.strictPartial, offline, latencyMs, snap.fail, ftp.fail, ftp.delayMs, rtsp.refuse, rtsp.reset]
        count: { type: integer, minimum: 1, description: on for the next N matching requests only }
        ms: { type: integer, minimum: 0 }
        cmds: { type: array, items: { type: string } }
        rspCode: { type: integer }
    SeedClip:
      type: object
      required: [daysAgo, start, end, triggers]
      properties:
        daysAgo: { type: integer, minimum: 0, maximum: 365 }
        start: { type: string, pattern: '^\d{6}$', description: HHMMSS camera-local }
        end: { type: string, pattern: '^\d{6}$' }
        mainEnd: { type: string, pattern: '^\d{6}$' }
        triggers: { type: array, items: { type: string, enum: [motion, person, vehicle, pet] } }
security:
  - bearer: []
paths:
  /healthz:
    get:
      summary: Liveness and readiness (no auth, no details)
      security: []
      responses:
        '200': { description: up }
  /sim/login:
    post:
      summary: Web UI login - exchanges the control token for a session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [token], properties: { token: { type: string } } }
      responses:
        '204': { description: 'logged in; Set-Cookie camsim_session_<name> (name lowercased, non-alphanumerics removed; HttpOnly, SameSite=Strict, 12 h)' }
        '401': { description: wrong token }
  /sim/logout:
    post:
      summary: Web UI logout - clears the session cookie
      security: []
      responses:
        '204': { description: logged out }
  /sim/session:
    get:
      summary: Whether the request carries a valid session cookie
      security: []
      responses:
        '200': { description: '{ loggedIn: boolean }' }
  /sim/api/state:
    get:
      summary: 'Identity, faults, recent events, SD card, counters, running settings and the SD pipeline ({on:false[, error]} or {on:true, until, running})'
      responses:
        '200': { description: state }
        '401': { description: missing or wrong token }
  /sim/api/videos:
    get:
      summary: The video library and the selected video
      description: >
        `{ selected, videos: [{ id, name, state, converted, durationS, error }] }`.
        `test-pattern` is always there; the rest come from CAMSIM_LIBRARY_DIR and are
        prepared one at a time after start (state pending → preparing → ready or failed).
      responses:
        '200': { description: library }
  /sim/api/video:
    put:
      summary: Select the video the camera shows
      description: >
        Live (FLV and RTSP), snapshots and new recordings switch to it. Recordings
        made earlier keep their own video.
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [id], properties: { id: { type: string } } }
      responses:
        '200': { description: '{ selected }' }
        '400': { description: invalid id }
        '404': { description: unknown video }
        '409': { description: video not ready (still preparing, or failed) }
  /sim/api/videos/{id}/poster:
    get:
      summary: A still of a ready video (JPEG)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: image/jpeg }
        '404': { description: unknown or not ready }
  /sim/api/events:
    get:
      summary: Recent events, newest first
      parameters:
        - { name: limit, in: query, schema: { type: integer, maximum: 500 } }
      responses:
        '200': { description: events }
    post:
      summary: Trigger an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, durationS]
              properties:
                type: { type: string, enum: [motion, person, vehicle, pet] }
                durationS: { type: integer, minimum: 1, maximum: 3600 }
      responses:
        '201': { description: the event and the recording it starts or extends (null when not recorded) }
        '409': { description: the camera is powered off }
        '400': { description: invalid input, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /sim/api/recordings/seed:
    post:
      summary: Add past recordings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clips:
                  oneOf:
                    - { type: string, enum: [demo] }
                    - { type: array, items: { $ref: '#/components/schemas/SeedClip' } }
      responses:
        '201': { description: added }
        '400': { description: invalid input }
  /sim/api/recordings:
    get:
      summary: Recordings of one day
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, pattern: '^\\d{4}-\\d{2}-\\d{2}$' } }
      responses:
        '200': { description: recordings }
        '400': { description: invalid date }
    delete:
      summary: Empty the SD card
      responses:
        '204': { description: emptied }
  /sim/api/pipeline:
    post:
      summary: Switch the SD pipeline on for `minutes` (1 to CAMSIM_PIPELINE_MAX_MIN, default 60); it switches itself off then
      description: >
        Re-encodes the live SD stream (FLV channel0_sub, RTSP h264Preview_01_sub) with the camera's
        name, date and time, watermark (Osd) and flip/mirror (Isp.rotation, Isp.mirroring). Main,
        snapshots, recordings and downloads are unchanged. Never persisted: off after a restart or reset.
        Calling it while on sets a new end time.
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { minutes: { type: integer, minimum: 1, maximum: 1440 } } }
      responses:
        '200': { description: '{on:true, until:<unix ms>, running:<bool>}' }
        '400': { description: 'invalid minutes: {error: invalid, detail}' }
    delete:
      summary: Switch the SD pipeline off
      responses:
        '204': { description: off }
  /sim/api/faults:
    get:
      summary: Active faults
      responses:
        '200': { description: faults, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Fault' } } } } }
    delete:
      summary: Clear all faults
      responses:
        '204': { description: cleared }
  /sim/api/faults/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string } }
    put:
      summary: Switch a fault on (body holds its parameters)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Fault' }
      responses:
        '200': { description: the active fault }
        '400': { description: unknown fault or missing parameter }
    delete:
      summary: Switch a fault off
      responses:
        '204': { description: cleared }
  /sim/api/actions/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, enum: [tokens.revoke, reboot, power-off, power-on, flv.dropActive, downloads.dropActive, clear, factory-reset] } }
    post:
      summary: >
        One-shot action. reboot takes {ms, dropsConnection}; power-on takes {ms} (boot time).
        power-off drops every connection and ends sessions until power-on.
        clear empties recordings, the event list and counters (settings, certificate and sessions stay).
        factory-reset returns settings, faults, video, certificate and content to the factory state,
        drops waiting FTP uploads, then reboots (takes {ms}); 409 while not on.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ms: { type: integer, minimum: 0, maximum: 600000 }
                dropsConnection: { type: boolean }
      responses:
        '202': { description: reboot or power-on started }
        '204': { description: done }
        '400': { description: unknown action or bad parameter }
        '409': { description: 'not possible in this power state (powered_off, already_off, already_on, or busy: power-off while booting, factory-reset while rebooting)' }
  /sim/api/reset:
    post:
      summary: Back to a known state; an empty body resets everything
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                settings: { type: boolean }
                recordings: { type: boolean }
                counters: { type: boolean }
                faults: { type: boolean }
                video: { type: boolean, description: back to the test pattern }
      responses:
        '204': { description: reset }
  /sim/api/media/snapshot:
    get:
      summary: A snapshot JPEG (for the web UI; not a camera request)
      responses:
        '200': { description: image/jpeg }
  /sim/api/media/live/{stream}:
    parameters:
      - { name: stream, in: path, required: true, schema: { type: string, enum: [sub, main] } }
    get:
      summary: Endless live FLV (main is H.265 with codec id 12); not counted as a camera stream
      responses:
        '200': { description: video/x-flv }
        '404': { description: unknown stream }
  /sim/api/recordings/days:
    get:
      summary: Days with recordings in a month (one character per day)
      parameters:
        - { name: year, in: query, required: true, schema: { type: integer } }
        - { name: mon, in: query, required: true, schema: { type: integer, minimum: 1, maximum: 12 } }
      responses:
        '200': { description: '{year, mon, table}' }
        '400': { description: invalid input }
  /sim/api/recordings/{id}/{stream}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
      - { name: stream, in: path, required: true, schema: { type: string, enum: [sub, main] } }
    get:
      summary: A recording's file (video/mp4 with Content-Length; ?download=1 adds Content-Disposition)
      responses:
        '200': { description: video/mp4 }
        '404': { description: unknown recording or stream }
  /sim/api/settings:
    get:
      summary: Running settings plus devInfo, hddInfo, enc and certificate
      responses:
        '200': { description: settings }
  /sim/api/settings/{key}:
    parameters:
      - { name: key, in: path, required: true, schema: { type: string, enum: [Rec, MdAlarm, Isp, IrLights, WhiteLed, Osd, NetPort, Ftp] } }
    put:
      summary: Write the WHOLE settings object, with the camera's validation
      description: >
        Osd (camera name, date and time, the `watermark` logo) is stored, validated and
        returned like the others, but nothing is drawn on the video: cam-sim plays its
        source (the test pattern or a library video) as it is.
      responses:
        '200': { description: the stored object }
        '400': { description: 'invalid value: {error: invalid, rspCode}' }
        '404': { description: unknown key }
  /sim/api/settings/AiAlarm/{type}:
    parameters:
      - { name: type, in: path, required: true, schema: { type: string, enum: [people, vehicle, dog_cat] } }
    put:
      summary: Write the whole AiAlarm object for one AI type
      responses:
        '200': { description: the stored object }
        '400': { description: invalid value }
        '404': { description: unknown type }
  /sim/api/users:
    get:
      summary: Camera users (names and levels, no passwords)
      responses:
        '200': { description: users }
  /sim/api/requests:
    get:
      summary: Recent camera API requests (no query strings, no secrets)
      parameters:
        - { name: limit, in: query, schema: { type: integer, maximum: 500 } }
      responses:
        '200': { description: requests }
  /sim/api/stream:
    get:
      summary: Server-Sent Events (state, event, request, fault, ftp, video; the certificate arrives as a state event)
      responses:
        '200': { description: text/event-stream }